backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
python-multipart==0.0.6

# Database & ORM
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
geoalchemy2

python-dotenv
shapely
tweepy

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Real-time & Caching
redis==5.0.1
celery==5.3.4
websockets==11.0.3

# AI/ML Libraries
transformers
torch
pillow==10.0.1
opencv-python-headless
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.24.3

# Geospatial
geoalchemy2==0.14.2
shapely==2.0.2

# API Integration
requests
httpx==0.25.1
aiofiles==23.2.0

# Development
pytest==7.4.3
black==23.10.1
flake8==6.1.0
python-dotenv==1.0.0